requirements.txt
# Django Core
Django==5.0.0
python-decouple==3.8

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4

# Celery & Redis
celery==5.3.4
redis==5.0.1

# AI/ML - Updated versions for compatibility
torch==2.1.2
transformers==4.47.1
sentence-transformers==3.3.1

# CrewAI & Groq
crewai==0.203.1
groq==0.4.1

# Web Scraping & RSS
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2

# Audio Processing
openai-whisper==20231117
soundfile==0.12.1

# Data Processing
numpy==1.26.2
pandas==2.1.4
python-dateutil==2.8.2